Give the procfs reading its own module
The Linux inspection grew past the file-size gate carrying two jobs:
deciding what counts as a stable, matching identity, and reading and
parsing what the kernel publishes. The procfs half — the process
source, the executable observation, and the stat, pid, and user-id
parsers — moves to its own module; the inspection logic keeps the
parent. Code motion only; no behavior change.
